Fix production build
The main issue revolves around the fact that some part of the build process (electron-builder) prevents the “build” folder to be packaged within the app.asar file. The way this has been fixed is to rename “build” to “bundle” - this should be revisited at some stage, because the fact this cannot be changed is kinda ridiculous… I don’t understand the obsession with preventing configuration to create-react-app, there literally isn’t any reason to prevent developers from configuring their toolchain.